Proxy/requirements.txt
pyserial
paho-mqtt
pyyaml